api/microservices/games/requirements.txt
grpc-interceptor ~= 0.11.0
grpcio-tools ~= 1.30
pytest ~= 5.4
pymongo ~= 3.11.2
dnspython ~= 2.1.0